// fe/src/lib/api/generated/model/myProfileResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export interface MyProfileResponse {
  civil_status: string;
  contact_number: string;
  current_address: string;
  date_of_birth: string | null;
  email: string;
  first_name: string;
  full_name: string;
  institutional_id: string | null;
  last_name: string;
  middle_name: string;
  permanent_address: string;
  profile_photo_updated_at: string | null;
  profile_photo_url: string | null;
  role: string;
  suffix: string;
  user_id: string;
}